added Caption Validation
